+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [9.10.0](https://github.com/javalent/initiative-tracker/compare/9.9.0...9.10.0) (2023-05-02)
+
+
+### Features
+
+* You can now create custom encounter builder headers ([2c42744](https://github.com/javalent/initiative-tracker/commit/2c42744b3ec0b20059d5ff68cdc6ec037415effa))
+
+
+### Bug Fixes
+
+* can disable encounter builder ribbon icon ([d6d2bee](https://github.com/javalent/initiative-tracker/commit/d6d2bee38d423ac1512ec81b5704f3a540839b17))
+* can now open combatant view for creatures in encounter builder ([d6d2bee](https://github.com/javalent/initiative-tracker/commit/d6d2bee38d423ac1512ec81b5704f3a540839b17))
+* creatures in table display difficulty ([921034a](https://github.com/javalent/initiative-tracker/commit/921034a695a04312ffa09768329e9fdc791b8e04))
+* encounter builder name is editable ([d6d2bee](https://github.com/javalent/initiative-tracker/commit/d6d2bee38d423ac1512ec81b5704f3a540839b17))
+* launching encounter from builder works again ([d6d2bee](https://github.com/javalent/initiative-tracker/commit/d6d2bee38d423ac1512ec81b5704f3a540839b17))
+* new encounter builder ribbon icon ([d6d2bee](https://github.com/javalent/initiative-tracker/commit/d6d2bee38d423ac1512ec81b5704f3a540839b17))
+* player & xp sections of builder are collapsible ([8505873](https://github.com/javalent/initiative-tracker/commit/85058739e04b9947fe08dc93e732bf3095dbf5b1))
+
+## [9.9.0](https://github.com/javalent/initiative-tracker/compare/9.8.2...9.9.0) (2023-05-02)
+
+
+### Features
+
+* encounter creatures can be clicked to open combatant view ([6fea212](https://github.com/javalent/initiative-tracker/commit/6fea2126d68857847d4e7a837371ecf3c979a7e0))
+
+
+### Bug Fixes
+
+* `friendly` can be used for allies (close [#157](https://github.com/javalent/initiative-tracker/issues/157)) ([114896f](https://github.com/javalent/initiative-tracker/commit/114896f91250338dea4036e0696e6fe799abc3f8))
+* Add to Encounter will add players that aren't already in the encounter (close [#160](https://github.com/javalent/initiative-tracker/issues/160)) ([37a6520](https://github.com/javalent/initiative-tracker/commit/37a6520bc10d4b06022371a176ee200ce8656df9))
+* display hidden & friendly icons in encounter ([b46545f](https://github.com/javalent/initiative-tracker/commit/b46545fbf7faea19775a5f0b70def400045c00e5))
+* hp state for rollable creatures now persists ([b8fb407](https://github.com/javalent/initiative-tracker/commit/b8fb407b46614e218821b2d48d79a48ed73c92aa))
+* save initiative state when resetting ac ([f362df1](https://github.com/javalent/initiative-tracker/commit/f362df1a5c8f7903665458d79d5bcee8acbf7873))
+
 ## [9.8.2](https://github.com/javalent/initiative-tracker/compare/9.8.1...9.8.2) (2023-04-25)